moved online-editor-page to challenges-module
