+++ frontend/src/app/modules/challenges/online-editor-page/online-editor-page.component.sass
